+++ starkbank/transfer/__transfer.py
@@ -1,6 +1,6 @@
 from ..utils import rest
-from ..utils.checks import check_datetime, check_date, check_datetime_or_date
-from ..utils.resource import Resource
+from starkcore.utils.checks import check_datetime, check_date, check_datetime_or_date
+from starkcore.utils.resource import Resource
 
 
 class Transfer(Resource):