+++ starkbank/error.py
@@ -1,0 +1,1 @@
+from starkcore.error import InputErrors, Error, StarkError, UnknownError, InternalServerError, InvalidSignatureError