+++ starkbank/key.py
@@ -1,0 +1,1 @@
+from starkcore.key import create