+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.15.0"
+version = "2.16.0"
 
 user = None
 language = "en-US"